Connection to swad api working
