+++ SWADroid/src/main/java/es/ugr/swad/swadroid/modules/location/ManageLocation.java
@@ -6,29 +6,22 @@ import android.content.Intent;
 import android.content.IntentFilter;
 import android.net.wifi.ScanResult;
 import android.net.wifi.WifiManager;
+import android.os.AsyncTask;
 import android.os.Bundle;
 import android.util.Log;
 import android.view.View;
 import android.widget.ArrayAdapter;
 import android.widget.ListView;
-import android.widget.Switch;
-import android.widget.TextView;
 import android.widget.Toast;
 
-import com.android.volley.Request;
-import com.android.volley.RequestQueue;
-import com.android.volley.Response;
-import com.android.volley.VolleyError;
-import com.android.volley.toolbox.JsonObjectRequest;
-import com.android.volley.toolbox.Volley;
 import com.google.android.material.floatingactionbutton.FloatingActionButton;
 import com.journeyapps.barcodescanner.BarcodeEncoder;
 
-import org.json.JSONException;
-import org.json.JSONObject;
+import org.ksoap2.serialization.SoapObject;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Vector;
 import java.util.concurrent.Executors;
 import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.ScheduledFuture;
@@ -38,6 +31,7 @@ import es.ugr.swad.swadroid.Constants;
 import es.ugr.swad.swadroid.R;
 import es.ugr.swad.swadroid.gui.MenuActivity;
 import es.ugr.swad.swadroid.model.Location;
+import es.ugr.swad.swadroid.model.Pair;
 import es.ugr.swad.swadroid.model.UserFilter;
 import es.ugr.swad.swadroid.modules.login.Login;
 import es.ugr.swad.swadroid.modules.messages.SearchUsers;
@@ -50,53 +44,18 @@ public class ManageLocation extends MenuActivity {
      * Messages tag name for Logcat
      */
     private static final String TAG = Constants.APP_TAG + " Manage location";
-
-    /**
-     * Code width
-     */
-    private static final int CODE_WIDTH = 250;
-
-    /**
-     * Code height
-     */
-    private static final int CODE_HEIGHT = 250;
-
-    /**
-     * Barcode Encoder
-     */
-    private static BarcodeEncoder barcodeEncoder;
-    /**
-     * Find user location text&icon
-     */
-    private static FloatingActionButton findUser;
     /**
      * Location history
      */
-    private static ListView history;
-    /**
-     * Allow to share location
-     */
-    private static Boolean allowSharingLocation=false;
-    /**
-     * Update location
-     */
-    private static FloatingActionButton updateLocation;
+    private ListView history;
     /**
      * Manage wifi properties
      */
     private WifiManager wifiManager;
     /**
-     * Arraylisto to store SSID of available networks
+     * Arraylist to store SSID of available networks
      */
-    private ArrayList<String> availableNetworks = new ArrayList<>();
-    /**
-     * List to store ScanResults of available networks
-     */
-    private List<ScanResult> results;
-    /**
-     * URL of MacStore
-     */
-    private String url = "https://mac-store.herokuapp.com/";
+    private List<Pair> availableNetworks = new ArrayList<>();
     /**
      * Adapter of LocationDistance array
      */
@@ -105,14 +64,6 @@ public class ManageLocation extends MenuActivity {
      * Array of receivers
      */
     private ArrayList<UserFilter> arrayReceivers;
-    /**
-     * Message's receivers (nicknames)
-     */
-    private String receivers;
-    /**
-     * Save if the list is expanded
-     */
-    private boolean showAll;
     private ScheduledExecutorService scheduler = Executors.newScheduledThreadPool(1);
     private static String METHOD_NAME = "";
     private static IWebserviceClient webserviceClient;
@@ -127,7 +78,6 @@ public class ManageLocation extends MenuActivity {
 
         setTitle(R.string.manageLocation);
         getSupportActionBar().setDisplayHomeAsUpEnabled(true);
-        barcodeEncoder = new BarcodeEncoder();
 
         history = findViewById(R.id.location_history_data);
         wifiManager = (WifiManager) getApplicationContext().getSystemService(Context.WIFI_SERVICE);
@@ -149,7 +99,10 @@ public class ManageLocation extends MenuActivity {
         super.onStart();
 
         //Find user location listener
-        findUser = findViewById(R.id.find_user);
+        /**
+         * Find user location text&icon
+         */
+        FloatingActionButton findUser = findViewById(R.id.find_user);
         findUser.setOnClickListener(new View.OnClickListener() {
            @Override
            public void onClick(View view) {
@@ -164,7 +117,10 @@ public class ManageLocation extends MenuActivity {
         //Get history
         history = findViewById(R.id.location_history_data);
 
-        updateLocation = findViewById(R.id.update_location);
+        /**
+         * Update location
+         */
+        FloatingActionButton updateLocation = findViewById(R.id.update_location);
         updateLocation.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
@@ -193,11 +149,18 @@ public class ManageLocation extends MenuActivity {
     protected void onActivityResult(int requestCode, int resultCode, Intent data) {
         if (data != null) {
             arrayReceivers = (ArrayList) data.getSerializableExtra("receivers");
-            receivers = "";
+            /**
+             * Message's receivers (nicknames)
+             */
+            String receivers = "";
             for(int i=0; i<arrayReceivers.size(); i++){
                 receivers += arrayReceivers.get(i) + ",";
             }
 
+            /**
+             * Save if the list is expanded
+             */
+            boolean showAll;
             if(arrayReceivers.size() <= 3)
                 showAll = false;
         }
@@ -215,15 +178,18 @@ public class ManageLocation extends MenuActivity {
         @Override
         public void onReceive(Context context, Intent intent) {
             Log.d(TAG, "onReceive se está ejecuntado");
-            results = wifiManager.getScanResults();
+            /**
+             * List to store ScanResults of available networks
+             */
+            List<ScanResult> results = wifiManager.getScanResults();
             unregisterReceiver(this);
             if (results.size() > 0) {
                 for (ScanResult scanResult : results) {
                     String bssid = scanResult.BSSID.replace(":", "");
                     double distance = (int) calculateDistance(scanResult.level, scanResult.frequency);
-                    //addLocations(bssid, distance);
                     try {
-                        getLocation(bssid);
+                        AsyncLocation asyncLocation = new AsyncLocation(bssid, distance);
+                        asyncLocation.execute();
                     } catch (Exception e) {
                         Log.d(TAG, "EXCEPCION AL OBTENER LA LOCALIZACIÓN");
                         e.printStackTrace();
@@ -295,12 +261,68 @@ public class ManageLocation extends MenuActivity {
      * @param simple Flag for select simple or complex response
      * @throws Exception
      */
-    private static void sendRequest(Class<?> cl, boolean simple) throws Exception {
+    protected List<Location> sendRequest(Class<?> cl, boolean simple) throws Exception {
         ((SOAPClient) webserviceClient).sendRequest(cl, simple);
         result = webserviceClient.getResult();
+
+        List<Location> locationList = new ArrayList<>();
+        if (result!=null) {
+
+            ArrayList<?> res = new ArrayList<>((Vector<?>) result);
+            SoapObject soap = (SoapObject) res.get(1);
+            int numLocations = soap.getPropertyCount();
+
+            for (int i = 0; i < numLocations; i++) {
+                SoapObject pii = (SoapObject) soap.getProperty(i);
+                Integer institutionCode = Integer.valueOf(pii.getProperty("institutionCode").toString());
+                String institutionShortName = pii.getProperty("institutionShortName").toString();
+                String institutionFullName = pii.getProperty("institutionShortName").toString();
+                Integer centerCode = Integer.valueOf(pii.getProperty("centerCode").toString());
+                String centerShortName = pii.getProperty("centerShortName").toString();
+                String centerFullName = pii.getProperty("centerFullName").toString();
+                Integer buildingCode = Integer.valueOf(pii.getProperty("buildingCode").toString());
+                String buildingShortName = pii.getProperty("buildingShortName").toString();
+                String buildingFullName = pii.getProperty("buildingFullName").toString();
+                Integer floor = Integer.valueOf(pii.getProperty("floor").toString());
+                Integer roomCode = Integer.valueOf(pii.getProperty("roomCode").toString());
+                String roomShortName = pii.getProperty("roomShortName").toString();
+                String roomFullName = pii.getProperty("roomFullName").toString();
+
+                locationList.add(new Location(institutionCode, institutionShortName, institutionFullName,
+                        centerCode, centerShortName, centerFullName, buildingCode, buildingShortName, buildingFullName,
+                        floor, roomCode, roomShortName, roomFullName));
+            }
+        }
+        return locationList;
     }
 
-    private static void getLocation(String mac) throws Exception {
+    private class AsyncLocation extends AsyncTask<String, Void, String> {
+
+        private String mac;
+        private Double distance;
+
+        AsyncLocation(String mac, Double distance){
+            super();
+            this.mac = mac;
+            this.distance = distance;
+        }
+
+        @Override
+        protected String doInBackground(String... strings) {
+            try {
+                for (Location l : getLocation(mac)) {
+                    Pair<Location, Double> locationInfo = new Pair<>(l, distance);
+                    availableNetworks.add(locationInfo);
+                    adapter.notifyDataSetChanged();
+                }
+            } catch (Exception e) {
+                e.printStackTrace();
+            }
+            return null;
+        }
+    }
+
+    private List<Location> getLocation(String mac) throws Exception {
         Log.d(TAG, "Get location from MAC address");
 
         METHOD_NAME= "getLocations";
@@ -308,6 +330,6 @@ public class ManageLocation extends MenuActivity {
         createRequest(SOAPClient.CLIENT_TYPE);
         addParam("wsKey", Login.getLoggedUser().getWsKey());
         addParam("MAC", mac);
-        sendRequest(Object.class, false);
+        return sendRequest(Object.class, false);
     }
 }
